second commit

+++ lms_app/lms_view/apply_view/views.py
@@ -5,28 +5,35 @@ from django.http import HttpRequest
 from django.shortcuts import render, redirect
 
 from lms_app.lms_dto.ApplyDto import ApplyDto
-from lms_app.models import Comment, Apply
+from lms_app.models import Apply
 from lms_app.service_controllers import service_controller, CommentDto
 
 
 @login_required(redirect_field_name='next')
 def apply_for_a_course(request, course_id):
+    if request.user.has_perm('lms_app.add_apply'):
 
-    l_as_list = []
-    for g in request.user.groups.all():
-        l_as_list.append(g.name)
+        l_as_list = []
+        for g in request.user.groups.all():
+            l_as_list.append(g.name)
 
-    username = request.user.username
-    user_id = request.user.id
-    tutor_id = service_controller.tutor_management_service().details(user_id).id
+        username = request.user.username
+        user_id = request.user.id
+        tutor_id = service_controller.tutor_management_service().details(user_id).id
 
-    context = {
-        'username': username,
-        'l_as_list': l_as_list,
-    }
-    __create_if_post_method(request, course_id, tutor_id, context)
-    if request.method == 'POST':
-        return redirect('tutor_details')
+        context = {
+            'username': username,
+            'l_as_list': l_as_list,
+        }
+        __create_if_post_method(request, course_id, tutor_id, context)
+        if request.method == 'POST':
+            return redirect('tutor_details')
+
+    else:
+        context={
+            'message': 'You are not authorised'
+        }
+        return render(request, 'error_message.html', context)
 
 def edit_application(request, apply_id):
     pass
@@ -34,12 +41,18 @@ def edit_application(request, apply_id):
 
 @login_required(redirect_field_name='next')
 def cancel_application(request, apply_id):
-    try:
-        service_controller.apply_management_service().delete(apply_id)
-        return redirect('tutor_details')
-    except Apply.DoesNotExist as e:
-        print('This application does not exist!')
-        raise e
+    if request.user.is_superuser:
+        try:
+            service_controller.apply_management_service().delete(apply_id)
+            return redirect('tutor_details')
+        except Apply.DoesNotExist as e:
+            print('This application does not exist!')
+            raise e
+    else:
+        context={
+            'message': 'You are not authorised'
+        }
+        return render(request, 'error_message.html', context)
 
 
 def __application_attribute_request(request: HttpRequest, course_id, tutor_id):
